refactor(cli): Fix CodeQL py/overwritten-inherited-attribute warnings
Use an Interface instead of Inheritance, and move the `cli_parser` attribute inside the classes because the types are too different.
The SubgroupConfigParsingManager expects a SubgroupConfigParser, and RootConfigParsingManager expects a RootConfigParser.
Even if both inherit from BaseConfigParser, their calling conventions differ too much, and it raises warnings in Pycharm and CodeQL.