fix(actor): Enable ZMQ High-Water-Mark on actors sockets
Setting the HWM value to 0 is a very bad idea because it breaks the
backpressure in the pipeline. For example, if the formula is too slow
to handle reports, the puller will not stop fetching data and a huge
backlog of reports will stay in memory, leading to a linear increase in
the memory usage of PowerAPI. This problem is most likely to happen when
using the puller in post-mortem mode.