!feat(formula): Use report type to pusher actor mapping for actor
Change the `pushers` argument of the `FormulaActor` class from a dict
mapping the actor name to a pusher actor to a dict mapping the report
type to a list of pusher actors.

This allow to have a more efficient dispatch of the generated reports to
theirs corresponding pushers in the formula report handlers.